# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bot Marketplace API (MVP)
  version: 0.1.0
  license:
    name: Proprietary
    url: ../LICENSE.md
  description: >
    MVP API for strategies, bots, and signals (demo-first).

servers:
  - url: /api

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      operationId: healthzGet
      summary: Liveness probe
      security: []
      responses:
        "200":
          description: OK
        "503":
          description: Service Unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /readyz:
    get:
      operationId: readyzGet
      summary: Readiness probe
      security: []
      responses:
        "200":
          description: OK
        "503":
          description: Service Unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /auth/login:
    post:
      operationId: authLogin
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/Problem"

  /strategies:
    get:
      operationId: strategiesList
      summary: List strategies
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      operationId: strategiesCreate
      summary: Create strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Strategy"
      responses:
        "201":
          description: Created
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"

  /strategies/{strategyId}:validate:
    post:
      operationId: strategiesValidate
      summary: Validate a strategy payload without saving
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Strategy"
      responses:
        "200":
          description: Valid
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"

  /bots:
    get:
      operationId: botsList
      summary: List bots
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      operationId: botsCreate
      summary: Create bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotCreateRequest"
      responses:
        "201":
          description: Created
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"

  /bots/{botId}/events:
    get:
      operationId: botEventsList
      summary: Bot transition/event log
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"

  # ── Exchange Connections (Stage 8) ─────────────────────────────────────────

  /exchanges:
    get:
      operationId: exchangesList
      summary: List exchange connections (no secrets returned)
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExchangeConnectionView"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      operationId: exchangesCreate
      summary: Create exchange connection
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeConnectionCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeConnectionView"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"

  /exchanges/{connectionId}:
    get:
      operationId: exchangesGet
      summary: Get exchange connection (no secrets returned)
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeConnectionView"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: exchangesPatch
      summary: Update exchange connection (re-encrypts secret if provided)
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeConnectionPatchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeConnectionView"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: exchangesDelete
      summary: Delete exchange connection
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"

  /exchanges/{connectionId}/test:
    post:
      operationId: exchangesTest
      summary: Test exchange connection (demo-first)
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeConnectionTestResult"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"

  /signals/webhook/{strategyId}:
    post:
      operationId: signalWebhook
      summary: Webhook signal
      security: []
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalRequest"
      responses:
        "202":
          description: Accepted
        "400":
          $ref: "#/components/responses/Problem"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  parameters:
    XWorkspaceId:
      name: X-Workspace-Id
      in: header
      required: true
      schema:
        type: string
      description: Workspace ID (membership is enforced server-side)

  responses:
    Problem:
      description: Problem Details (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }

    LoginResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }

    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }

    Strategy:
      $ref: "../schema/strategy.schema.json"

    BotCreateRequest:
      type: object
      additionalProperties: false
      required: [strategyId]
      properties:
        strategyId: { type: string }
        note: { type: string }

    ExchangeConnectionView:
      type: object
      description: >
        Safe view of an exchange connection — never includes apiKey, secret, or encryptedSecret.
      additionalProperties: false
      required: [id, workspaceId, exchange, name, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        exchange: { type: string, example: "BYBIT" }
        name: { type: string, example: "Main Bybit account" }
        status:
          type: string
          enum: [UNKNOWN, CONNECTED, FAILED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    ExchangeConnectionCreateRequest:
      type: object
      additionalProperties: false
      required: [exchange, name, apiKey, secret]
      properties:
        exchange: { type: string, example: "BYBIT" }
        name: { type: string, example: "Main Bybit account" }
        apiKey: { type: string, minLength: 1 }
        secret: { type: string, minLength: 1, description: "Raw secret — stored encrypted, never returned" }

    ExchangeConnectionPatchRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        apiKey: { type: string, minLength: 1 }
        secret: { type: string, minLength: 1, description: "Raw secret — re-encrypted on update, never returned" }

    ExchangeConnectionTestResult:
      type: object
      additionalProperties: false
      required: [id, status, detail]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [CONNECTED, FAILED]
        detail: { type: string }

    SignalRequest:
      type: object
      additionalProperties: false
      required: [side, symbol, ts]
      properties:
        side:
          type: string
          enum: [Buy, Sell]
        symbol:
          type: string
        reason:
          type: string
        ts:
          type: string
          format: date-time